feat: Dedupes equal queries fired at the same time
Several components might fire the same query at the same time. In this
case, the fetchPolicy is not used and both queries are executed. Here,
we use the queryDefinition as a key in a cache that holds pending
queries. If an existing query for a particular query definition is
already pending, its promise is returned to the caller.